+++ packages/project-management-automation/lib/tasks/assign-fixed-issues/index.js
@@ -1,4 +1,4 @@
-const debug = require( '../../debug' );
+import debug from '../../debug.js';
 
 /** @typedef {ReturnType<typeof import('@actions/github').getOctokit>} GitHub */
 /** @typedef {import('@octokit/webhooks-types').EventPayloadMap['pull_request']} WebhookPayloadPullRequest */
@@ -44,4 +44,4 @@ async function assignFixedIssues( payload, octokit ) {
 	}
 }
 
-module.exports = assignFixedIssues;
+export default assignFixedIssues;